origin: exclude server-launching metadata e2e tests on Windows
Match the repo convention (metrics_e2e, tpc_fed, storage_metrics) by
tagging the subprocess-launching and full InitializeHandlers e2e suites
//go:build !windows; move shared param-setter helpers to an untagged
file so the Windows-running admin-auth test still compiles.
